+++ zio-connect/shared/src/test/scala/zio/connect/file/LiveFileConnectorSpec.scala
@@ -7,7 +7,7 @@ object LiveFileConnectorSpec extends FileConnectorSpec {
   override def spec =
     suite("LiveFileConnectorSpec")(fileConnectorSpec)
       .provideSome[Scope](
-        FileOps.live,
+        LiveFileOps.layer,
         LiveFileConnector.layer
       )
