+++ zio-connect/shared/src/test/scala/zio/connect/file/testkit/TestFileConnectorSpec.scala
@@ -2,7 +2,6 @@ package zio.connect.file.testkit
 
 import com.google.common.jimfs.{Configuration, Jimfs}
 import zio.connect.file.FileConnectorSpec
-import zio.connect.testkit.file.TestFileConnector
 import zio.{Scope, ZLayer}
 
 object TestFileConnectorSpec extends FileConnectorSpec {
@@ -12,7 +11,7 @@ object TestFileConnectorSpec extends FileConnectorSpec {
       .provideSome[Scope](
         ZLayer.succeed(Jimfs.newFileSystem(Configuration.forCurrentPlatform())),
         TestFileOps.inMemory,
-        TestFileConnector.layer
+        TestFileConnector.layerWithCustomFileSystem
       )
 
 }
